Add files via upload
Almost have serial code integrated, but in SerialCommsViewModel method ListAvailablePorts(), it's not working for some reason and there is a comment detailing what's going on. Ideally it would be good to connect to the device before clicking the "Enter" button so that's what the ConnectToTracer() method is for and why it's called from MainPage(), but currently it's not working so it wouldn't let the app deploy/execute successfully.