Add files via upload
Slightly modified functionality of clear button so it clears the data if there was anything, if the textbox was empty it now returns it the home screen which matches the clients typical behaviour. Also adjusted the spacing of the 7, 8, 9, and return buttons so the space above them wasn't larger than the rest of the spacing for the grid. The code for connecting to the serial device is implemented in the return Btn handler, but commented out so nothing breaks when you try to use it.