Add files via upload
Moved Device class back to it's own file so it's cleaner. 
Changed the write method to write byte by byte instead of the job Id as a string
I am still somehow unable to read data though which is very strange.
Removed a lot of unnecessary included libraries in the different files, and spaced out and formatted things nicely with a few new comments after looking at the example Object Design document.