Add files via upload
Formatted code a bit more and added matching formatted comments at the top of each file to follow for future files. Also updated method names to all be ThisIsAMethod style and class properties to be _thisIsAClassProperty so they are diffarentiable from any local varibles. Fixed a couple of the compile warnings that were occurring with accidental conversions from double to int as well.