Add files via upload
I made quite a few changes this weekend, the crosshair on the lens turns green when it's within plus/minus 5 pixels to the location, the job number sends on return and to the PC like always and parses the received location given a random format I sent it in which is just 000xy000, just to make sure I can parse the received data, where x and y are three digit coordinates. From there it saves the x and y position and places the crosshair at that location. I changed the enter job number screen so that they can only enter a 6 digit number, (if job number == 6 then backspace one digit and set last digit to the one previously entered). Session history list box is populated with time stamp and job numbers entered for that session. Settings button initialized the feed and opens the settings flyout. And the config file is parsed and updates the sliders upon first initialization of the feed. 
I'm going to tidy a few things up then I will push my code 